mercury/requirements.txt
django==3.2.5
djangorestframework==3.12.4
django-filter==21.1
markdown==3.3.6
celery==5.1.2
sqlalchemy==1.4.27
gevent==21.12.0
nbconvert==6.3.0
pyyaml==6.0
django-cors-headers==3.10.1
ipython==7.30.1
ipykernel==6.6.0 
psutil==5.8.0
whitenoise==5.3.0
python-dotenv==0.19.2